+++ packages/hardhat/hardhat.config.js
@@ -5,8 +5,9 @@ const ethers = require('ethers');
 require('@nomiclabs/hardhat-ethers');
 require('solidity-coverage');
 
-require('./tasks/deploy')
-require('./tasks/sale-state')
+require('./tasks/deploy');
+require('./tasks/sale-state');
+require('./tasks/mint');
 
 module.exports = {
   solidity: '0.8.4',
