+++ packages/hardhat/package.json
@@ -36,6 +36,7 @@
     "@nomiclabs/hardhat-ethers": "2.0.2",
     "@openzeppelin/contracts": "4.3.0",
     "del": "6.0.0",
+    "enquirer": "2.3.6",
     "prompt-list": "3.2.0",
     "solidity-coverage": "0.7.17"
   }
